+++ core/settlement/engine.go
@@ -154,10 +154,6 @@ func (e *Engine) AddTrade(trade *types.Trade) {
 	e.mu.Lock()
 	defer e.mu.Unlock()
 	// network registers a wash trade to update its position
-	if trade.Buyer == types.NetworkParty && trade.Buyer == trade.Seller {
-		e.addNetworkTrade(trade)
-		return
-	}
 	var buyerSize, sellerSize int64
 	// checking the len of cd shouldn't be required here, but it is needed in the second if
 	// in case the buyer and seller are one and the same...
@@ -195,29 +191,6 @@ func (e *Engine) AddTrade(trade *types.Trade) {
 		size:        -size,
 		newSize:     sellerSize - size,
 	})
-}
-
-func (e *Engine) addNetworkTrade(trade *types.Trade) {
-	tSize := int64(trade.Size)
-	// sell wash trade, the settled position of the network decreases.
-	if trade.Aggressor == types.SideSell {
-		tSize *= -1
-	}
-
-	// the trade happens at the current mark price, so it is considered a settled position
-	e.settledPosition[types.NetworkParty] += tSize
-	// any unsettled trades (though this slice should be empty) should be updated accordingly
-	trades, ok := e.trades[types.NetworkParty]
-	if !ok {
-		return
-	}
-	// meaning that the size and new size should be updated to include the wash trade
-	for i, t := range trades {
-		t.size += tSize
-		t.newSize += tSize
-		trades[i] = t
-	}
-	e.trades[types.NetworkParty] = trades
 }
 
 func (e *Engine) HasTraded() bool {
@@ -298,15 +271,15 @@ func (e *Engine) SettleMTM(ctx context.Context, markPrice *num.Uint, positions [
 	e.trades = map[string][]*settlementTrade{} // remove here, once we've processed it all here, we're done
 	evts := make([]events.Event, 0, len(positions))
 	var (
-		largestShare  *mtmTransfer       // pointer to whomever gets the last remaining amount from the loss
-		zeroShares    = []*mtmTransfer{} // all zero shares for equal distribution if possible
-		zeroAmts      = false
-		mtmDec        = num.NewDecimalFromFloat(0)
-		lossTotal     = num.UintZero()
-		winTotal      = num.UintZero()
-		lossTotalDec  = num.NewDecimalFromFloat(0)
-		winTotalDec   = num.NewDecimalFromFloat(0)
-		appendLargest = false
+		largestShare *mtmTransfer       // pointer to whomever gets the last remaining amount from the loss
+		zeroShares   = []*mtmTransfer{} // all zero shares for equal distribution if possible
+		zeroAmts     = false
+		mtmDec       = num.NewDecimalFromFloat(0)
+		lossTotal    = num.UintZero()
+		winTotal     = num.UintZero()
+		lossTotalDec = num.NewDecimalFromFloat(0)
+		winTotalDec  = num.NewDecimalFromFloat(0)
+		// appendLargest = false
 	)
 
 	// network is treated as a regular party
@@ -375,21 +348,21 @@ func (e *Engine) SettleMTM(ctx context.Context, markPrice *num.Uint, positions [
 	e.mu.Unlock()
 	delta := num.UintZero().Sub(lossTotal, winTotal)
 	// make sure largests share is never nil
-	if largestShare == nil {
-		largestShare = &mtmTransfer{
-			MarketPosition: &npos{
-				price: markPrice.Clone(),
-			},
-		}
-		appendLargest = true
-	}
+	/*
+		if largestShare == nil {
+			largestShare = &mtmTransfer{
+				MarketPosition: &npos{
+					price: markPrice.Clone(),
+				},
+			}
+			appendLargest = true
+		}*/
 	if !delta.IsZero() {
 		if zeroAmts {
-			if appendLargest {
-				zeroShares = append(zeroShares, largestShare)
-			}
+			// if appendLargest {
+			// zeroShares = append(zeroShares, largestShare)
+			// }
 			zRound := num.DecimalFromInt64(int64(len(zeroShares)))
-			zeroShares = append(zeroShares, largestShare)
 			// there are more transfers from losses than we pay out to wins, but some winning parties have zero transfers
 			// this delta should == combined win decimals, let's sanity check this!
 			if winTotalDec.LessThan(lossTotalDec) && winTotalDec.LessThan(lossTotalDec.Sub(zRound)) {
@@ -420,9 +393,9 @@ func (e *Engine) SettleMTM(ctx context.Context, markPrice *num.Uint, positions [
 	}
 	// append wins after loss transfers
 	transfers = append(transfers, wins...)
-	if len(transfers) > 0 && appendLargest && largestShare.transfer != nil {
-		transfers = append(transfers, largestShare)
-	}
+	// if len(transfers) > 0 && appendLargest && largestShare.transfer != nil {
+	// transfers = append(transfers, largestShare)
+	// }
 	if len(evts) > 0 {
 		e.broker.SendBatch(evts)
 	}